chore: organize run-v5 tests into unit and integration
