+++ packages/run-v5/test/integration/logs.integration.test.js
@@ -4,9 +4,8 @@
 const cmd = require('../../commands/logs')
 const cli = require('heroku-cli-util')
 const {expect} = require('chai')
-const {describeAcceptance} = require('../test-helper')
 
-describeAcceptance('logs', () => {
+describe('logs', () => {
   beforeEach(() => cli.mockConsole())
 
   it('shows the logs', () => {
